+++ src/Fable.Transforms/Python/Replacements.fs
@@ -1779,7 +1779,12 @@ let arrayModule (com: ICompiler) (ctx: Context) r (t: Type) (i: CallInfo) (_: Ex
 
     match i.CompiledName, args with
     | "ToSeq", [ arg ] -> Some arg
-    | "OfSeq", [ arg ] -> toArray r t arg |> Some
+    | "OfSeq", [ arg ] ->
+        let meth = Naming.lowerFirst i.CompiledName
+        let args = injectArg com ctx r "Array" meth i.GenericArgs args
+
+        Helper.LibCall(com, "array", meth, t, args, i.SignatureArgTypes, ?loc = r)
+        |> Some
     | "OfList", [ arg ] ->
         Helper.LibCall(com, "list", "to_array", t, args, i.SignatureArgTypes, ?loc = r)
         |> Some